pyproject.toml
[tool.poetry]
name = "qzemoji"
version = "2.1.1a1"
description = "Translate Qzone emoji link to text."
authors = ["JamzumSum <zzzzss990315@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/JamzumSum/QzEmoji"
repository = "https://github.com/JamzumSum/QzEmoji"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.32"}
aiosqlite = "^0.17.0"
assetsupdater = {git = "https://github.com/JamzumSum/AssetsUpdater.git", branch = "release", extras=["async"]}
PyYaml = {version = "^6.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-asyncio = "^0.18.2"
pre-commit = "^2.17.0"
black = "^22.1.0"
isort = "^5.10.1"

[tool.poetry.extras]
build = ["PyYaml"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# customize begin
[tool.pytest.ini_options]
pythonpath = 'src'
asyncio_mode = 'strict'
log_cli = 1
log_cli_level = 'WARNING'


[tool.isort]
profile = "google"

[tool.black]
line-length = 99
target-version = ['py37']
